+++ components/nfc/tools/impl/src/main/java/com/flipperdevices/nfc/tools/impl/api/NfcToolsApiImpl.kt
@@ -2,28 +2,26 @@ package com.flipperdevices.nfc.tools.impl.api
 
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.core.log.LogTagProvider
-import com.flipperdevices.core.log.info
+import com.flipperdevices.nfc.tools.api.MfKey32Nonce
 import com.flipperdevices.nfc.tools.api.NfcToolsApi
 import com.flipperdevices.nfc.tools.impl.bindings.MfKey32Binding
 import com.squareup.anvil.annotations.ContributesBinding
-import java.lang.Long.parseLong
 import javax.inject.Inject
 
 @ContributesBinding(AppGraph::class, NfcToolsApi::class)
 class NfcToolsApiImpl @Inject constructor() : NfcToolsApi, LogTagProvider {
     override val TAG = "NfcToolsApi"
 
-    @Suppress("MagicNumber")
-    override fun test() {
+    override suspend fun bruteforceKey(mfKey32Nonce: MfKey32Nonce): ULong? {
         val key = MfKey32Binding.tryRecoverKey(
-            uid = parseLong("2a234f80", 16),
-            nt0 = parseLong("55721809", 16),
-            nr0 = parseLong("ce9985f6", 16),
-            ar0 = parseLong("772f55be", 16),
-            nt1 = parseLong("a27173f2", 16),
-            nr1 = parseLong("e386b505", 16),
-            ar1 = parseLong("5fa65203", 16)
+            uid = mfKey32Nonce.uid.toLong(),
+            nt0 = mfKey32Nonce.nt0.toLong(),
+            nr0 = mfKey32Nonce.nr0.toLong(),
+            ar0 = mfKey32Nonce.ar0.toLong(),
+            nt1 = mfKey32Nonce.nt1.toLong(),
+            nr1 = mfKey32Nonce.nr1.toLong(),
+            ar1 = mfKey32Nonce.ar1.toLong()
         )
-        info { "Read key $key" }
+        return key?.toULongOrNull()
     }
 }